+++ noodles-cram/src/writer/header_container.rs
@@ -3,16 +3,22 @@ mod header;
 use std::io::{self, Write};
 
 use bytes::BufMut;
+use flate2::Compression;
 use noodles_sam as sam;
 
 use self::header::write_header;
 use super::container::write_block;
-use crate::container::{block::ContentType, Block};
+use crate::{
+    codecs::Encoder,
+    container::{block::ContentType, Block},
+};
 
 pub fn write_header_container<W>(writer: &mut W, header: &sam::Header) -> io::Result<()>
 where
     W: Write,
 {
+    const ENCODER: Encoder = Encoder::Gzip(Compression::new(6));
+
     validate_reference_sequences(header.reference_sequences())?;
 
     let header_data = header.to_string().into_bytes();
@@ -25,8 +31,7 @@ where
 
     let block = Block::builder()
         .set_content_type(ContentType::FileHeader)
-        .set_uncompressed_len(data.len())
-        .set_data(data.into())
+        .compress_and_set_data(data, ENCODER)?
         .build();
 
     write_header(writer, block.len())?;
@@ -55,24 +60,49 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_write_header_container() -> io::Result<()> {
+    fn test_write_header_container() -> Result<(), Box<dyn std::error::Error>> {
+        use byteorder::{LittleEndian, WriteBytesExt};
+        use flate2::CrcWriter;
         use sam::header::record::value::{
             map::{self, header::Version},
             Map,
         };
 
+        use crate::{codecs::gzip, writer::num::write_itf8};
+
         let header_header = Map::<map::Header>::new(Version::new(1, 6));
         let header = sam::Header::builder().set_header(header_header).build();
 
-        let mut buf = Vec::new();
-        write_header_container(&mut buf, &header)?;
+        let mut actual = Vec::new();
+        write_header_container(&mut actual, &header)?;
 
-        let mut expected = Vec::new();
+        let header_data = b"@HD\tVN:1.6\n";
+        let header_data_len = i32::try_from(header_data.len())?;
 
-        // header
-        expected.extend_from_slice(&[
-            0x18, 0x00, 0x00, 0x00, // length = 24
-            0xff, 0xff, 0xff, 0xff, 0x0f, // reference sequence ID = -1 (None)
+        let mut data = Vec::new();
+        data.put_i32_le(header_data_len);
+        data.extend(header_data);
+
+        let compressed_data = gzip::encode(Compression::new(6), &data)?;
+
+        let mut block_writer = CrcWriter::new(Vec::new());
+        block_writer.write_all(&[
+            0x01, // compression method = 1 (Gzip)
+            0x00, // content type = 0 (FileHeader)
+            0x00, // block content ID = 0
+        ])?;
+        write_itf8(&mut block_writer, i32::try_from(compressed_data.len())?)?; // compressed length
+        write_itf8(&mut block_writer, i32::try_from(data.len())?)?; // uncompressed length
+        block_writer.write_all(&compressed_data)?;
+
+        let crc32 = block_writer.crc().sum();
+        let mut expected_block = block_writer.into_inner();
+        expected_block.put_u32_le(crc32); // crc32
+
+        let mut header_writer = CrcWriter::new(Vec::new());
+        header_writer.write_i32::<LittleEndian>(i32::try_from(expected_block.len())?)?; // length
+        write_itf8(&mut header_writer, -1)?; // reference sequence ID = -1 (None)
+        header_writer.write_all(&[
             0x00, // alignment start = 0
             0x00, // alignment span = 0
             0x00, // record count = 0
@@ -80,24 +110,16 @@ mod tests {
             0x00, // base count = 0
             0x01, // block count = 1
             0x00, // landmarks.len = 0
-            0xf9, 0xf0, 0x9e, 0x3d, // CRC32 = 3d9ef0f9
-        ]);
+        ])?;
 
-        // block
-        expected.extend_from_slice(&[
-            0x00, // compression method = 0 (None)
-            0x00, // content type = 0 (FileHeader)
-            0x00, // block content ID = 0
-            0x0f, // compressed length = 15
-            0x0f, // uncompressed length = 15
-        ]);
-        expected.put_i32_le(11);
-        expected.extend_from_slice(b"@HD\tVN:1.6\n");
-        expected.extend_from_slice(&[
-            0xf4, 0xe5, 0x16, 0xd3, // CRC32 = d316e5f4
-        ]);
+        let crc32 = header_writer.crc().sum();
+        let mut expected_header = header_writer.into_inner();
+        expected_header.put_u32_le(crc32); // crc32
 
-        assert_eq!(buf, expected);
+        let mut expected = expected_header;
+        expected.extend(expected_block);
+
+        assert_eq!(actual, expected);
 
         Ok(())
     }